Rename LocalPlaybookText field
